改动很多，不想一一列举了，反正变的更牛逼了！

+++ natsublog-api/src/main/java/com/natsu/blog/mapper/TagMapper.java
@@ -9,6 +9,6 @@ import java.util.List;
 @Repository
 public interface TagMapper extends BaseMapper<Tag> {
 
-    List<Tag> getTagsByArticleId(int articleId);
+    List<Tag> getTagsByArticleId(Long articleId);
 
 }
